refactor: split aerospace category validation rules
